Made drawing of cells to the screen possible.
Added a Tileset class.

Added a default tileset image (media/tileset.png).

Added a class derived from Gosu::Window.

Added a System module to add a layer of abstraction over the Gosu
library. In particular, added the possibility to pass in drawing and
update procs which will be called by the window class. The drawing
routines will still be Gosu's for now.

Added #draw method to Cell to draw a cell to the screen.